Extract LB structured data functionality into common
