+++ article/app/views/liveBlog.scala.html
@@ -1,3 +1,4 @@
+@import model.liveblog.LiveBlogPage
 @(model: LiveBlogPage)(implicit request: RequestHeader, context: _root_.model.ApplicationContext)
 
 @import common.LinkTo
