+++ common/app/model/liveblog/LiveBlogPage.scala
@@ -1,0 +1,5 @@
+package model.liveblog
+
+import model.{Article, RelatedContent}
+
+case class LiveBlogPage(article: Article, currentPage: LiveBlogCurrentPage, related: RelatedContent) extends PageWithStoryPackage
