+++ common/app/model/liveblog/PageWithStoryPackage.scala
@@ -1,0 +1,10 @@
+package model.liveblog
+
+import model.{Article, ArticleSchemas, ContentPage, RelatedContent}
+
+trait PageWithStoryPackage extends ContentPage {
+  def article: Article
+  def related: RelatedContent
+  override lazy val item = article
+  val articleSchemas = ArticleSchemas
+}
