+++ common/app/views/support/structuredData/Image.scala
