+++ common/app/views/support/structuredData/LiveBlogPosting.scala
@@ -1,11 +1,16 @@
 package views.support.structuredData
 
 import common.LinkTo
-import controllers.LiveBlogPage
+import model.Cached.WithoutRevalidationResult
+import model.{Article, Cached}
+import model.liveblog._
+import model.liveblog.ParseBlockId.{InvalidFormat, ParsedBlockId}
 import org.joda.time.DateTime
 import play.api.libs.json._
 import play.api.mvc.RequestHeader
 import views.support.Format
+
+import scala.concurrent.Future
 
 // Since Json-Ld has fields that start with the @ character I can't just marshall it from
 // case classes which would be simpler. I've opted to construct the JsonValues manually. I
@@ -15,23 +20,21 @@ object LiveBlogPosting {
 
   def zulu(date: DateTime)(implicit request: RequestHeader): String = Format(date, "yyyy-MM-dd'T'HH:mm:ssZ")
 
-  def apply(blog: LiveBlogPage)(implicit request: RequestHeader): JsValue = {
-
-    val blocks = blog.currentPage.currentPage.blocks
+  def apply(blog: Article, blocks: Seq[BodyBlock])(implicit request: RequestHeader): JsValue = {
 
     Json.obj(
       "@context" -> "http://schema.org",
       "@type" -> "LiveBlogPosting",
-      "url" -> LinkTo(blog.article.metadata.url),
-      "headline" -> blog.article.trail.headline,
-      "description" -> blog.article.fields.trailText,
-      "datePublished" -> zulu(blog.article.trail.webPublicationDate),
-      "coverageStartTime" -> zulu(blog.article.trail.webPublicationDate),
-      "coverageEndTime" -> zulu(blog.article.fields.lastModified),
-      "dateModified" -> zulu(blog.article.fields.lastModified),
+      "url" -> LinkTo(blog.metadata.url),
+      "headline" -> blog.trail.headline,
+      "description" -> blog.fields.trailText,
+      "datePublished" -> zulu(blog.trail.webPublicationDate),
+      "coverageStartTime" -> zulu(blog.trail.webPublicationDate),
+      "coverageEndTime" -> zulu(blog.fields.lastModified),
+      "dateModified" -> zulu(blog.fields.lastModified),
       "publisher" -> Organisation(),
       "liveBlogUpdate" -> Json.arr(blocks.map(
-        block => BlogPosting(blog, block)
+       block => BlogPosting(blog, block)
       ))
     ) ++ MainMedia(blog).getOrElse(Json.obj())
